finquest_backend/requirements.txt
fastapi[standard]
uvicorn
python-dotenv
mysql-connector-python
PyJWT
requests
openai